# appui/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "AuthSessionResponse": {
        "description": "Bearer token plus user identity returned after register/login.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserResponse"
          }
        },
        "required": [
          "access_token",
          "user"
        ],
        "title": "AuthSessionResponse",
        "type": "object"
      },
      "AuthUserResponse": {
        "description": "Authenticated user identity returned to the frontend.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ],
        "title": "AuthUserResponse",
        "type": "object"
      },
      "BacktestJobStatusResponse": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u5f02\u6b65\u56de\u6d4b\u4efb\u52a1\u72b6\u6001\u67e5\u8be2\u54cd\u5e94\u3002",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "BacktestJobStatusResponse",
        "type": "object"
      },
      "BacktestJobSubmitResponse": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u5f02\u6b65\u56de\u6d4b\u63d0\u4ea4\u540e\u7684\u4efb\u52a1\u53e5\u67c4\u3002",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "BacktestJobSubmitResponse",
        "type": "object"
      },
      "BacktestRunRequest": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u524d\u7aef\u53d1\u8d77\u540c\u6b65\u6216\u5f02\u6b65\u56de\u6d4b\u65f6\u63d0\u4ea4\u7684\u8bf7\u6c42\u3002",
        "properties": {
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "title": "Metrics",
            "type": "array"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "symbol": {
            "examples": [
              "RB0909"
            ],
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "BacktestRunRequest",
        "type": "object"
      },
      "BacktestRunResponse": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u4e00\u6b21\u56de\u6d4b\u5b8c\u6210\u540e\u7684\u5b8c\u6574\u524d\u7aef\u54cd\u5e94\u3002",
        "properties": {
          "equity_curve": {
            "items": {
              "$ref": "#/components/schemas/EquityPoint"
            },
            "title": "Equity Curve",
            "type": "array"
          },
          "final_equity": {
            "title": "Final Equity",
            "type": "number"
          },
          "initial_cash": {
            "title": "Initial Cash",
            "type": "number"
          },
          "metrics": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Metrics",
            "type": "object"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/BacktestTrade"
            },
            "title": "Trades",
            "type": "array"
          }
        },
        "required": [
          "symbol",
          "strategy",
          "initial_cash",
          "final_equity",
          "trades",
          "equity_curve",
          "metrics"
        ],
        "title": "BacktestRunResponse",
        "type": "object"
      },
      "BacktestTrade": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u56de\u6d4b\u6210\u4ea4\u660e\u7ec6\uff0c\u7528\u4e8e\u524d\u7aef\u5c55\u793a\u4e70\u5356\u70b9\u3002",
        "properties": {
          "cash": {
            "title": "Cash",
            "type": "number"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "side": {
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "integer"
          }
        },
        "required": [
          "time",
          "side",
          "price",
          "quantity",
          "cash",
          "reason"
        ],
        "title": "BacktestTrade",
        "type": "object"
      },
      "Candle": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u4e00\u6839\u6807\u51c6 OHLCV \u8721\u70db\u3002",
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "time": {
            "description": "Unix timestamp in seconds",
            "title": "Time",
            "type": "integer"
          },
          "volume": {
            "title": "Volume",
            "type": "number"
          }
        },
        "required": [
          "time",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "Candle",
        "type": "object"
      },
      "EquityPoint": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u5355\u4e2a\u6743\u76ca\u66f2\u7ebf\u70b9\uff0c\u7528\u4e8e\u5c55\u793a\u8d44\u4ea7\u3001\u73b0\u91d1\u548c\u6301\u4ed3\u4ef7\u503c\u53d8\u5316\u3002",
        "properties": {
          "cash": {
            "title": "Cash",
            "type": "number"
          },
          "equity": {
            "title": "Equity",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "position_value": {
            "title": "Position Value",
            "type": "number"
          },
          "time": {
            "title": "Time",
            "type": "integer"
          }
        },
        "required": [
          "time",
          "equity",
          "cash",
          "position_value",
          "position"
        ],
        "title": "EquityPoint",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KLineResponse": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u5e26\u5206\u9875\u4fe1\u606f\u7684\u4e00\u7ec4\u5408\u7ea6 K \u7ebf\u6570\u636e\u3002",
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/Candle"
            },
            "title": "Candles",
            "type": "array"
          },
          "limit": {
            "description": "Maximum candle count requested",
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Zero-based offset of this response window",
            "title": "Offset",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "total": {
            "description": "Total candles available for the symbol",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "symbol",
          "total",
          "offset",
          "limit",
          "candles"
        ],
        "title": "KLineResponse",
        "type": "object"
      },
      "LoginRequest": {
        "additionalProperties": false,
        "description": "Email login request.",
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MetricInfo": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u63cf\u8ff0\u4e00\u4e2a\u53ef\u5c55\u793a\u3001\u53ef\u9009\u62e9\u7684\u7ee9\u6548\u6307\u6807\u3002",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "MetricInfo",
        "type": "object"
      },
      "RegisterRequest": {
        "additionalProperties": false,
        "description": "Email-only registration request.",
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "StrategyInfo": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u63cf\u8ff0\u4e00\u4e2a\u53ef\u5c55\u793a\u3001\u53ef\u9009\u62e9\u7684\u56de\u6d4b\u7b56\u7565\u3002",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "StrategyInfo",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Futures Quantitative Research API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "login_user_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login User",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "get_me_api_auth_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "register_user_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register User",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/backtest/jobs": {
      "post": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u63d0\u4ea4\u5f02\u6b65\u56de\u6d4b\u4efb\u52a1\uff0c\u907f\u514d\u957f\u8017\u65f6\u8bf7\u6c42\u963b\u585e HTTP \u8c03\u7528\u3002",
        "operationId": "post_submit_backtest_job_api_backtest_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestJobSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Submit Backtest Job",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/backtest/jobs/{job_id}": {
      "get": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u67e5\u8be2\u5f02\u6b65\u56de\u6d4b\u4efb\u52a1\u7684\u5f53\u524d\u72b6\u6001\u548c\u5931\u8d25\u539f\u56e0\u3002",
        "operationId": "get_backtest_job_api_backtest_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestJobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Backtest Job",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/backtest/jobs/{job_id}/result": {
      "get": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u83b7\u53d6\u5df2\u5b8c\u6210\u5f02\u6b65\u56de\u6d4b\u4efb\u52a1\u7684\u9886\u57df\u7ed3\u679c\u3002",
        "operationId": "get_backtest_job_result_endpoint_api_backtest_jobs__job_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Backtest Job Result Endpoint",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/backtest/metrics": {
      "get": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u67e5\u8be2 quant_runtime \u5f53\u524d\u652f\u6301\u7684\u7ee9\u6548\u6307\u6807\u5143\u6570\u636e\u3002",
        "operationId": "list_metrics_api_backtest_metrics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MetricInfo"
                  },
                  "title": "Response List Metrics Api Backtest Metrics Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Metrics",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/backtest/run": {
      "post": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u540c\u6b65\u6267\u884c\u4e00\u6b21\u56de\u6d4b\u5e76\u76f4\u63a5\u8fd4\u56de\u5b8c\u6574\u7ed3\u679c\u3002",
        "operationId": "post_run_backtest_api_backtest_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Run Backtest",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/backtest/strategies": {
      "get": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u67e5\u8be2 quant_runtime \u5f53\u524d\u6ce8\u518c\u7684\u56de\u6d4b\u7b56\u7565\u5143\u6570\u636e\u3002",
        "operationId": "list_strategies_api_backtest_strategies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategyInfo"
                  },
                  "title": "Response List Strategies Api Backtest Strategies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Strategies",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/backtest/symbols": {
      "get": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u67e5\u8be2\u53ef\u7528\u4e8e\u56de\u6d4b\u7684 1 \u5206\u949f parquet \u5408\u7ea6\u5217\u8868\u3002",
        "operationId": "get_symbols_api_backtest_symbols_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Symbols Api Backtest Symbols Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Symbols",
        "tags": [
          "backtest"
        ]
      }
    },
    "/api/market/kline": {
      "get": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u8fd4\u56de\u5355\u4e2a\u5408\u7ea6\u7684\u4e00\u9875\u6807\u51c6 OHLCV K \u7ebf\u6570\u636e\u3002",
        "operationId": "get_kline_api_market_kline_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "examples": [
                "RB0909"
              ],
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 2000,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KLineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Kline",
        "tags": [
          "market"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "\u4e1a\u52a1\u529f\u80fd: \u7ed9\u90e8\u7f72\u3001\u4ee3\u7406\u548c\u672c\u5730\u8c03\u8bd5\u63d0\u4f9b\u8f7b\u91cf\u5b58\u6d3b\u68c0\u67e5\u3002",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}