# appui/src/api/generated/types.ts
// Generated from appapi OpenAPI schemas by scripts/generate_openapi.py.
// Do not edit this file by hand.

export interface AuthSessionResponse {
  access_token: string;
  token_type?: string;
  user: AuthUserResponse;
}

export interface AuthUserResponse {
  email: string;
  role: "admin" | "user";
}

export interface BacktestJobStatusResponse {
  job_id: string;
  status: "queued" | "running" | "succeeded" | "failed";
  error?: string | null;
}

export interface BacktestJobSubmitResponse {
  job_id: string;
  status: "queued" | "running" | "succeeded" | "failed";
}

export interface BacktestRunRequest {
  symbol: string;
  strategy?: string | null;
  start_time?: string | null;
  end_time?: string | null;
  metrics?: string[];
}

export interface BacktestRunResponse {
  symbol: string;
  strategy: string;
  initial_cash: number;
  final_equity: number;
  trades: BacktestTrade[];
  equity_curve: EquityPoint[];
  metrics: Record<string, number | null>;
}

export interface BacktestTrade {
  time: number;
  side: "buy" | "sell";
  price: number;
  quantity: number;
  cash: number;
  reason: string;
}

export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface EquityPoint {
  time: number;
  equity: number;
  cash: number;
  position_value: number;
  position: number;
}

export interface HTTPValidationError {
  detail?: ValidationError[];
}

export interface KLineResponse {
  symbol: string;
  total: number;
  offset: number;
  limit: number;
  candles: Candle[];
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface MetricInfo {
  id: string;
  name: string;
  description: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
}

export interface StrategyInfo {
  id: string;
  name: string;
  description: string;
}

export interface ValidationError {
  loc: (string | number)[];
  msg: string;
  type: string;
  input?: unknown;
  ctx?: Record<string, unknown>;
}

export type BacktestResult = BacktestRunResponse;
export type BacktestRunParams = BacktestRunRequest;